extract HellaswagMixin; drop samples_lock; spec threshold 1.8
