+++ test/registered/core/test_basic_sanity.py
@@ -1,6 +1,7 @@
-"""Basic sanity: small-but-broad server smoke that downstream stages
-depend on. Three sanity kits, one shared server, covering protocol
-contract, decode correctness, and scheduler stress paths."""
+"""Stage-a basic sanity: small-but-broad server smoke that downstream
+stages depend on. Multiple sanity-kit mixins driving one shared server,
+covering protocol, decode correctness, scheduler stress, occupancy, and
+hellaswag accuracy."""
 
 import unittest
 
@@ -10,6 +11,7 @@ from sglang.test.kits.basic_api_contract_kit import BasicAPIContractMixin
 from sglang.test.kits.basic_decode_correctness_kit import BasicDecodeCorrectnessMixin
 from sglang.test.kits.basic_scheduler_stress_kit import BasicSchedulerStressMixin
 from sglang.test.kits.fwd_occupancy_kit import FwdOccupancyMixin
+from sglang.test.kits.hellaswag_kit import HellaswagMixin
 from sglang.test.test_utils import (
     DEFAULT_MODEL_NAME_FOR_TEST,
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
@@ -27,6 +29,7 @@ class TestBasicSanity(
     BasicDecodeCorrectnessMixin,
     BasicSchedulerStressMixin,
     FwdOccupancyMixin,
+    HellaswagMixin,
     CustomTestCase,
 ):
     served_model_name = DEFAULT_MODEL_NAME_FOR_TEST
@@ -54,23 +57,6 @@ class TestBasicSanity(
     @classmethod
     def tearDownClass(cls):
         kill_process_tree(cls.process.pid)
-
-    def test_accuracy_floor(self):
-        # Stage-a's own accuracy gate -- catches systematic regressions
-        # that pass every cheap probe but tank multi-choice reasoning.
-        import sglang as sgl
-        from sglang.test.test_programs import test_hellaswag_select
-
-        sgl.set_default_backend(sgl.RuntimeEndpoint(self.base_url))
-        try:
-            accuracy, _ = test_hellaswag_select()
-        finally:
-            sgl.set_default_backend(None)
-        self.assertGreater(
-            accuracy,
-            0.60,
-            f"hellaswag accuracy floor breached: {accuracy:.3f}",
-        )
 
 
 if __name__ == "__main__":
